Add create and destroy same tx test
