+++ zk/realtime/test/constants.go
@@ -23,4 +23,7 @@ const (
 
 	destroyABIJson     = "[\n\t{\"inputs\":[{\"internalType\":\"address\",\"name\":\"_owner\",\"type\":\"address\"}],\"stateMutability\":\"payable\",\"type\":\"constructor\"},\n\t{\"inputs\":[{\"internalType\":\"address payable\",\"name\":\"recipient\",\"type\":\"address\"}],\"name\":\"destroy\",\"outputs\":[],\"stateMutability\":\"nonpayable\",\"type\":\"function\"},\n\t{\"inputs\":[],\"name\":\"owner\",\"outputs\":[{\"internalType\":\"address\",\"name\":\"\",\"type\":\"address\"}],\"stateMutability\":\"view\",\"type\":\"function\"}\n]"
 	destroyBytecodeStr = "60806040526040516103ac3803806103ac833981810160405281019061002591906100c8565b805f5f6101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff160217905550506100f3565b5f5ffd5b5f73ffffffffffffffffffffffffffffffffffffffff82169050919050565b5f6100978261006e565b9050919050565b6100a78161008d565b81146100b1575f5ffd5b50565b5f815190506100c28161009e565b92915050565b5f602082840312156100dd576100dc61006a565b5b5f6100ea848285016100b4565b91505092915050565b6102ac806101005f395ff3fe608060405234801561000f575f5ffd5b5060043610610033575f3560e01c8062f55d9d146100375780638da5cb5b14610053575b5f5ffd5b610051600480360381019061004c919061019a565b610071565b005b61005b610118565b60405161006891906101e5565b60405180910390f35b5f5f9054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff163373ffffffffffffffffffffffffffffffffffffffff16146100ff576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016100f690610258565b60405180910390fd5b8073ffffffffffffffffffffffffffffffffffffffff16ff5b5f5f9054906101000a900473ffffffffffffffffffffffffffffffffffffffff1681565b5f5ffd5b5f73ffffffffffffffffffffffffffffffffffffffff82169050919050565b5f61016982610140565b9050919050565b6101798161015f565b8114610183575f5ffd5b50565b5f8135905061019481610170565b92915050565b5f602082840312156101af576101ae61013c565b5b5f6101bc84828501610186565b91505092915050565b5f6101cf82610140565b9050919050565b6101df816101c5565b82525050565b5f6020820190506101f85f8301846101d6565b92915050565b5f82825260208201905092915050565b7f4e6f74206f776e657200000000000000000000000000000000000000000000005f82015250565b5f6102426009836101fe565b915061024d8261020e565b602082019050919050565b5f6020820190508181035f83015261026f81610236565b905091905056fea2646970667358221220d2ad35b6bbb9bb05e4c90b1c2b54e035735e1b0072a48b57c465e9698bfe555e64736f6c634300081d0033"
+
+	createDestroyABIJson     = "[\n\t{\"inputs\":[],\"stateMutability\":\"nonpayable\",\"type\":\"constructor\"},\n\t{\"anonymous\":false,\"inputs\":[{\"indexed\":false,\"internalType\":\"address\",\"name\":\"deployedAddress\",\"type\":\"address\"},{\"indexed\":false,\"internalType\":\"uint256\",\"name\":\"salt\",\"type\":\"uint256\"}],\"name\":\"ContractCreated\",\"type\":\"event\"},\n\t{\"anonymous\":false,\"inputs\":[{\"indexed\":false,\"internalType\":\"address\",\"name\":\"destroyedAddress\",\"type\":\"address\"},{\"indexed\":false,\"internalType\":\"address\",\"name\":\"recipient\",\"type\":\"address\"}],\"name\":\"ContractDestroyed\",\"type\":\"event\"},\n\t{\"inputs\":[{\"internalType\":\"bytes\",\"name\":\"bytecode\",\"type\":\"bytes\"},{\"internalType\":\"uint256\",\"name\":\"salt\",\"type\":\"uint256\"}],\"name\":\"computeAddress\",\"outputs\":[{\"internalType\":\"address\",\"name\":\"\",\"type\":\"address\"}],\"stateMutability\":\"view\",\"type\":\"function\"},\n\t{\"inputs\":[{\"internalType\":\"uint256\",\"name\":\"salt\",\"type\":\"uint256\"},{\"internalType\":\"address payable\",\"name\":\"recipient\",\"type\":\"address\"}],\"name\":\"createAndDestroy\",\"outputs\":[{\"internalType\":\"address\",\"name\":\"\",\"type\":\"address\"}],\"stateMutability\":\"payable\",\"type\":\"function\"},\n\t{\"inputs\":[{\"internalType\":\"uint256[]\",\"name\":\"salts\",\"type\":\"uint256[]\"},{\"internalType\":\"address payable\",\"name\":\"recipient\",\"type\":\"address\"}],\"name\":\"createMultipleAndDestroy\",\"outputs\":[{\"internalType\":\"address[]\",\"name\":\"\",\"type\":\"address[]\"}],\"stateMutability\":\"payable\",\"type\":\"function\"},\n\t{\"inputs\":[{\"internalType\":\"uint256\",\"name\":\"salt\",\"type\":\"uint256\"}],\"name\":\"createOnly\",\"outputs\":[{\"internalType\":\"address\",\"name\":\"\",\"type\":\"address\"}],\"stateMutability\":\"payable\",\"type\":\"function\"},\n\t{\"inputs\":[{\"internalType\":\"bytes\",\"name\":\"bytecode\",\"type\":\"bytes\"},{\"internalType\":\"uint256\",\"name\":\"salt\",\"type\":\"uint256\"}],\"name\":\"deployWithCreate2\",\"outputs\":[{\"internalType\":\"address\",\"name\":\"\",\"type\":\"address\"}],\"stateMutability\":\"payable\",\"type\":\"function\"},\n\t{\"inputs\":[{\"internalType\":\"address\",\"name\":\"contractAddress\",\"type\":\"address\"},{\"internalType\":\"address payable\",\"name\":\"recipient\",\"type\":\"address\"}],\"name\":\"destroyOnly\",\"outputs\":[],\"stateMutability\":\"nonpayable\",\"type\":\"function\"},\n\t{\"inputs\":[],\"name\":\"targetBytecode\",\"outputs\":[{\"internalType\":\"bytes\",\"name\":\"\",\"type\":\"bytes\"}],\"stateMutability\":\"view\",\"type\":\"function\"}\n]"
+	createDestroyBytecodeStr = "608060405234801561000f575f5ffd5b506040518060200161002090610047565b6020820181038252601f19601f820116604052505f90816100419190610291565b50610360565b610363806113d583390190565b5f81519050919050565b7f4e487b71000000000000000000000000000000000000000000000000000000005f52604160045260245ffd5b7f4e487b71000000000000000000000000000000000000000000000000000000005f52602260045260245ffd5b5f60028204905060018216806100cf57607f821691505b6020821081036100e2576100e161008b565b5b50919050565b5f819050815f5260205f209050919050565b5f6020601f8301049050919050565b5f82821b905092915050565b5f600883026101447fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff82610109565b61014e8683610109565b95508019841693508086168417925050509392505050565b5f819050919050565b5f819050919050565b5f61019261018d61018884610166565b61016f565b610166565b9050919050565b5f819050919050565b6101ab83610178565b6101bf6101b782610199565b848454610115565b825550505050565b5f5f905090565b6101d66101c7565b6101e18184846101a2565b505050565b5b81811015610204576101f95f826101ce565b6001810190506101e7565b5050565b601f8211156102495761021a816100e8565b610223846100fa565b81016020851015610232578190505b61024661023e856100fa565b8301826101e6565b50505b505050565b5f82821c905092915050565b5f6102695f198460080261024e565b1980831691505092915050565b5f610281838361025a565b9150826002028217905092915050565b61029a82610054565b67ffffffffffffffff8111156102b3576102b261005e565b5b6102bd82546100b8565b6102c8828285610208565b5f60209050601f8311600181146102f9575f84156102e7578287015190505b6102f18582610276565b865550610358565b601f198416610307866100e8565b5f5b8281101561032e57848901518255600182019150602085019450602081019050610309565b8683101561034b5784890151610347601f89168261025a565b8355505b6001600288020188555050505b505050505050565b6110688061036d5f395ff3fe60806040526004361061006f575f3560e01c8063933290dc1161004d578063933290dc1461010f578063b7b119ab1461013f578063ed16bf1a1461016f578063f3fd08ee146101975761006f565b806301df1822146100735780635ce7b0fc146100a357806381176c20146100d3575b5f5ffd5b61008d60048036038101906100889190610897565b6101c1565b60405161009a9190610901565b60405180910390f35b6100bd60048036038101906100b89190610a56565b610298565b6040516100ca9190610901565b60405180910390f35b3480156100de575f5ffd5b506100f960048036038101906100f49190610a56565b6102b9565b6040516101069190610901565b60405180910390f35b61012960048036038101906101249190610aeb565b610301565b6040516101369190610901565b60405180910390f35b61015960048036038101906101549190610b86565b610477565b6040516101669190610c9a565b60405180910390f35b34801561017a575f5ffd5b5061019560048036038101906101909190610ce4565b610726565b005b3480156101a2575f5ffd5b506101ab6107c8565b6040516101b89190610d82565b60405180910390f35b5f5f6102545f80546101d290610dcf565b80601f01602080910402602001604051908101604052809291908181526020018280546101fe90610dcf565b80156102495780601f1061022057610100808354040283529160200191610249565b820191905f5260205f20905b81548152906001019060200180831161022c57829003601f168201915b505050505084610298565b90507f1dc05c1d6a563dddb6c22082af72b54ec2f0207ceb55db5d13cdabc208f303a98184604051610287929190610e0e565b60405180910390a180915050919050565b5f5f8284516020860134f59050803b6102af575f5ffd5b8091505092915050565b5f5f60ff60f81b308486805190602001206040516020016102dd9493929190610f0e565b604051602081830303815290604052805190602001209050805f1c91505092915050565b5f5f6103945f805461031290610dcf565b80601f016020809104026020016040519081016040528092919081815260200182805461033e90610dcf565b80156103895780601f1061036057610100808354040283529160200191610389565b820191905f5260205f20905b81548152906001019060200180831161036c57829003601f168201915b505050505085610298565b90507f1dc05c1d6a563dddb6c22082af72b54ec2f0207ceb55db5d13cdabc208f303a981856040516103c7929190610e0e565b60405180910390a18073ffffffffffffffffffffffffffffffffffffffff1662f55d9d846040518263ffffffff1660e01b81526004016104079190610f6a565b5f604051808303815f87803b15801561041e575f5ffd5b505af1158015610430573d5f5f3e3d5ffd5b505050507f5cfb18adc50621d486741738e017b29b9f3932fa8741182bd791c1410b756b1a8184604051610465929190610fde565b60405180910390a18091505092915050565b60605f8484905067ffffffffffffffff81111561049757610496610932565b5b6040519080825280602002602001820160405280156104c55781602001602082028036833780820191505090505b5090505f5f90505b8585905081101561071a576105825f80546104e790610dcf565b80601f016020809104026020016040519081016040528092919081815260200182805461051390610dcf565b801561055e5780601f106105355761010080835404028352916020019161055e565b820191905f5260205f20905b81548152906001019060200180831161054157829003601f168201915b505050505087878481811061057657610575611005565b5b90506020020135610298565b82828151811061059557610594611005565b5b602002602001019073ffffffffffffffffffffffffffffffffffffffff16908173ffffffffffffffffffffffffffffffffffffffff16815250507f1dc05c1d6a563dddb6c22082af72b54ec2f0207ceb55db5d13cdabc208f303a982828151811061060357610602611005565b5b602002602001015187878481811061061e5761061d611005565b5b90506020020135604051610633929190610e0e565b60405180910390a181818151811061064e5761064d611005565b5b602002602001015173ffffffffffffffffffffffffffffffffffffffff1662f55d9d856040518263ffffffff1660e01b815260040161068d9190610f6a565b5f604051808303815f87803b1580156106a4575f5ffd5b505af11580156106b6573d5f5f3e3d5ffd5b505050507f5cfb18adc50621d486741738e017b29b9f3932fa8741182bd791c1410b756b1a8282815181106106ee576106ed611005565b5b602002602001015185604051610705929190610fde565b60405180910390a180806001019150506104cd565b50809150509392505050565b8173ffffffffffffffffffffffffffffffffffffffff1662f55d9d826040518263ffffffff1660e01b815260040161075e9190610f6a565b5f604051808303815f87803b158015610775575f5ffd5b505af1158015610787573d5f5f3e3d5ffd5b505050507f5cfb18adc50621d486741738e017b29b9f3932fa8741182bd791c1410b756b1a82826040516107bc929190610fde565b60405180910390a15050565b5f80546107d490610dcf565b80601f016020809104026020016040519081016040528092919081815260200182805461080090610dcf565b801561084b5780601f106108225761010080835404028352916020019161084b565b820191905f5260205f20905b81548152906001019060200180831161082e57829003601f168201915b505050505081565b5f604051905090565b5f5ffd5b5f5ffd5b5f819050919050565b61087681610864565b8114610880575f5ffd5b50565b5f813590506108918161086d565b92915050565b5f602082840312156108ac576108ab61085c565b5b5f6108b984828501610883565b91505092915050565b5f73ffffffffffffffffffffffffffffffffffffffff82169050919050565b5f6108eb826108c2565b9050919050565b6108fb816108e1565b82525050565b5f6020820190506109145f8301846108f2565b92915050565b5f5ffd5b5f5ffd5b5f601f19601f8301169050919050565b7f4e487b71000000000000000000000000000000000000000000000000000000005f52604160045260245ffd5b61096882610922565b810181811067ffffffffffffffff8211171561098757610986610932565b5b80604052505050565b5f610999610853565b90506109a5828261095f565b919050565b5f67ffffffffffffffff8211156109c4576109c3610932565b5b6109cd82610922565b9050602081019050919050565b828183375f83830152505050565b5f6109fa6109f5846109aa565b610990565b905082815260208101848484011115610a1657610a1561091e565b5b610a218482856109da565b509392505050565b5f82601f830112610a3d57610a3c61091a565b5b8135610a4d8482602086016109e8565b91505092915050565b5f5f60408385031215610a6c57610a6b61085c565b5b5f83013567ffffffffffffffff811115610a8957610a88610860565b5b610a9585828601610a29565b9250506020610aa685828601610883565b9150509250929050565b5f610aba826108c2565b9050919050565b610aca81610ab0565b8114610ad4575f5ffd5b50565b5f81359050610ae581610ac1565b92915050565b5f5f60408385031215610b0157610b0061085c565b5b5f610b0e85828601610883565b9250506020610b1f85828601610ad7565b9150509250929050565b5f5ffd5b5f5ffd5b5f5f83601f840112610b4657610b4561091a565b5b8235905067ffffffffffffffff811115610b6357610b62610b29565b5b602083019150836020820283011115610b7f57610b7e610b2d565b5b9250929050565b5f5f5f60408486031215610b9d57610b9c61085c565b5b5f84013567ffffffffffffffff811115610bba57610bb9610860565b5b610bc686828701610b31565b93509350506020610bd986828701610ad7565b9150509250925092565b5f81519050919050565b5f82825260208201905092915050565b5f819050602082019050919050565b610c15816108e1565b82525050565b5f610c268383610c0c565b60208301905092915050565b5f602082019050919050565b5f610c4882610be3565b610c528185610bed565b9350610c5d83610bfd565b805f5b83811015610c8d578151610c748882610c1b565b9750610c7f83610c32565b925050600181019050610c60565b5085935050505092915050565b5f6020820190508181035f830152610cb28184610c3e565b905092915050565b610cc3816108e1565b8114610ccd575f5ffd5b50565b5f81359050610cde81610cba565b92915050565b5f5f60408385031215610cfa57610cf961085c565b5b5f610d0785828601610cd0565b9250506020610d1885828601610ad7565b9150509250929050565b5f81519050919050565b5f82825260208201905092915050565b8281835e5f83830152505050565b5f610d5482610d22565b610d5e8185610d2c565b9350610d6e818560208601610d3c565b610d7781610922565b840191505092915050565b5f6020820190508181035f830152610d9a8184610d4a565b905092915050565b7f4e487b71000000000000000000000000000000000000000000000000000000005f52602260045260245ffd5b5f6002820490506001821680610de657607f821691505b602082108103610df957610df8610da2565b5b50919050565b610e0881610864565b82525050565b5f604082019050610e215f8301856108f2565b610e2e6020830184610dff565b9392505050565b5f7fff0000000000000000000000000000000000000000000000000000000000000082169050919050565b5f819050919050565b610e7a610e7582610e35565b610e60565b82525050565b5f8160601b9050919050565b5f610e9682610e80565b9050919050565b5f610ea782610e8c565b9050919050565b610ebf610eba826108e1565b610e9d565b82525050565b5f819050919050565b610edf610eda82610864565b610ec5565b82525050565b5f819050919050565b5f819050919050565b610f08610f0382610ee5565b610eee565b82525050565b5f610f198287610e69565b600182019150610f298286610eae565b601482019150610f398285610ece565b602082019150610f498284610ef7565b60208201915081905095945050505050565b610f6481610ab0565b82525050565b5f602082019050610f7d5f830184610f5b565b92915050565b5f819050919050565b5f610fa6610fa1610f9c846108c2565b610f83565b6108c2565b9050919050565b5f610fb782610f8c565b9050919050565b5f610fc882610fad565b9050919050565b610fd881610fbe565b82525050565b5f604082019050610ff15f8301856108f2565b610ffe6020830184610fcf565b9392505050565b7f4e487b71000000000000000000000000000000000000000000000000000000005f52603260045260245ffdfea26469706673582212203989bbd4fc6e82c436c3c9c8edb71890cdfc31ad82b5684e21ee065841a2bd0e64736f6c634300081d00336080604052335f5f6101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff1602179055503460018190555061030c806100575f395ff3fe608060405234801561000f575f5ffd5b506004361061003e575f3560e01c8062f55d9d146100425780633fa4f2451461005e5780638da5cb5b1461007c575b5f5ffd5b61005c600480360381019061005791906101c9565b61009a565b005b610066610141565b604051610073919061020c565b60405180910390f35b610084610147565b6040516100919190610245565b60405180910390f35b5f5f9054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff163373ffffffffffffffffffffffffffffffffffffffff1614610128576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161011f906102b8565b60405180910390fd5b8073ffffffffffffffffffffffffffffffffffffffff16ff5b60015481565b5f5f9054906101000a900473ffffffffffffffffffffffffffffffffffffffff1681565b5f5ffd5b5f73ffffffffffffffffffffffffffffffffffffffff82169050919050565b5f6101988261016f565b9050919050565b6101a88161018e565b81146101b2575f5ffd5b50565b5f813590506101c38161019f565b92915050565b5f602082840312156101de576101dd61016b565b5b5f6101eb848285016101b5565b91505092915050565b5f819050919050565b610206816101f4565b82525050565b5f60208201905061021f5f8301846101fd565b92915050565b5f61022f8261016f565b9050919050565b61023f81610225565b82525050565b5f6020820190506102585f830184610236565b92915050565b5f82825260208201905092915050565b7f4e6f74206f776e657200000000000000000000000000000000000000000000005f82015250565b5f6102a260098361025e565b91506102ad8261026e565b602082019050919050565b5f6020820190508181035f8301526102cf81610296565b905091905056fea26469706673582212208eb6a2eba50d5b33b8325e39ad8d213c20157e66b936613b6a314e409729c15264736f6c634300081d0033"
 )
